[Flexiglass] Fix FooterView flicker when collapsing any shade
When the shade collapses, there is a slight gap between finalizing the
fade out animation of the Footer, and hiding the SceneWindowRootView.
The problem is that after the shade expand/collapse has finished, the
FooterView is resetted to a default state with the  alpha of 1f. This is
causing a flicker.

Somehow the FooterView is the only child of the NSSL, that receives a
visible alpha in this state, so this CL brings it in parity with the
others.

Fixes: 439516450
Test: expand/collapse SingleShade -> look for FooterView flickers
Flag: com.android.systemui.scene_container
Change-Id: Ia64d78e507a96741ce296c99e50a6a1cc71515fc