Let Launcher call finish if cancel with snapshots
Do not call finishInner immediately upon cancel()
in RecentsController as a result of a transition merge
signal coming in for instance - if snapshots are sent to Launcher.

In this case, let Launcher handle the drawing of the snapshots,
and let it follow through with a "handshake", calling into
RecentsController to finish() on its own.

This CL also treats TRANSIT_PIP_BOUNDS_CHANGE same way as
TRANSIT_REMOVE_PIP upon merge() into RecentsController.
This way we also take snapshots when moving/resizing PiP.

Bug: 431716818
Flag: EXEMPT bugfix
Test: manually repro the steps in the bug
Change-Id: I372f1d43aa7f8b7521b3158a2baf074b6ff21ffa