Animate App Header to menu.
In order to fulfill the header -> menu open/close spec, the header needs
to be animating as if it shares the same surface as the first app info
pill of the menu. So we mimic this by having the pill's text and icon
exactly on top of each other, and then slowly scale it.

http://recall/-/hkw45bbkR06eCMwasLDKpi/yaSetG8VLujSHQMNHeIr4
http://recall/-/hkw45bbkR06eCMwasLDKpi/dJVQ7s36STMf6SSnj3MXqF

Bug: 314314136
Test: Manual, open/close the menu from the header.
Flag: com.android.window.flags.enable_drawing_app_handle

Change-Id: I85108b56bf000fcce59a1b96f167f0f161ba598b