Desktop screenshots: Temporary fix to ensure UI is excluded from image
Add a simple delay before taking a partial screenshot so that there is
time for the pre-capture UI to disappear before the region image is
captured. This is a temporary solution for the upcoming dogfood release
that should work for 99% of cases until a more robust solution is
implemented.

Bug: 435225255
Test: atest PreCapturViewModelTest
Test: atest ScreenshotInteractorTest
Flag: com.android.systemui.large_screen_screencapture
Change-Id: Ieb77eba277e4db52e056d9fa63c4cde749a1a0c3