Don't try to dequeue while shell transitions disabled
When systemui crashes, the transition system needs to
clean-up all pending/queued transitions. It seems that
some of the clean-up logic is "re-entrant" if the queue
is deep. This meant we were sometimes calling into
logic that acted as-if we weren't cleaning-up. So add
a check there.

Bug: 439829314
Test: N/A, this only happens during sysui crash -- watch crash trends
Flag: EXEMPT bugfix
Change-Id: I2ec9a76def7a1c77a0619d8810c5ecabc5da64f0