Exit split if task moves to bubble via trampoline
If a split screen task moves to bubble, via a trampoline, we only find
out about it when the transition animation starts.
This means we can't exit split screen as part of the same transition.

As split clears up the split screen state based on the TaskListener,
and this is triggered before the transition animation, bubbles can't
check that the task was previously in split.

Adding logic to the split StageTaskListener to detect cases when a task
is moving out of split screen and into a bubble. Exiting split screen in
such cases.

Bug: 440647718
Test: atest StageCoordinatorTests
Test: atest StageTaskListenerTests
Flag: com.android.wm.shell.fix_exit_split_on_enter_bubble
Change-Id: I47a97f08d50205b08d56972449b5cf0fc88eabdb