Updatable flashlight slider default level
We keep track of last user-set level after each slide and next time
the flashlight is enabled we turn on to that saved  level.

NOTE: The is a separation: when the flash is off, the flashlight repo
will say level is last enabled permanent level, but the CameraManager
will say level is default level. This is correct behavior, because at
this point if the user turns on the flashlight via repo it will turn
on to last enabled value, and if the user turns on the flashlight via
camera manager it  will turn on to default.

Bug: 438211714
Flag: com.android.systemui.flashlight_strength
Test: atest FlashlightRepositoryTest
Change-Id: I76f5642b6a03f46b01207726c48a73513aad160e