Tap and swipe API in relative display space.
When the display is created in the framework, the client doesn't
have a good signal about the display dimensions. They may have
changed since the last screenshot is taken. For max robustness,
convert from relative to pixel coordinates in the system server.

Also fixing the bug where the thread is blocked for 500ms during
a swipe. Now the swipe is scheduled in the background, no blocking
is done and the call returns immediately.

Also refactoring ComputerControlSessionImpl to make it a bit more
testable.

Bug: 428284603
Fix: 439774796
Test: atest & presubmit
Flag: android.companion.virtualdevice.flags.computer_control_access
Change-Id: I695e773363259b98430c4f77ba09d192fbfd7c3c