Split ImeOnBackInvokedDispatcher into two classes
ImeOnBackInvokedDispatcher was getting increasingly confusing because it
was shared between the app process and IME process. It contained some
logic that was only used at the app side and some logic that was only
used at the IME side.

This CL aims to simplify it by splitting the code into two new classes:
ImeBackCallbackProxy (app side) and ImeBackCallbackSender (IME side).
This makes it a lot clearer to readers what happens at the app side and
what happens at the IME side.

Bug: 438718906
Test: KeyboardVisibilityControlTest
Flag: EXEMPT low-risk refactor which would be difficult to flag
Change-Id: Id34d2170daf7144a5105d7fafb0507639d95e1f5