Remove wallpaper activity when display can't host tasks.
Fixes a bug where DesktopWallpaperActivity ends up on the internal
display due to mirroring transitions by removing it if the host display
can no longer host tasks. This occurs before core displayChange
handling, ensuring the wallpaper does not move to the front of the
internal display.

Note this is an incomplete fix: While this fixes the bad state reached
by the wallpaper being stuck in front, there is an animation bug where
the frontmost task on the internal display will flicker to the front of
the internal display, then animate offscreen. This is due to core
displayChange handling moving all tasks to front. This will require a
change to RootWindowContainer#onDisplayChanged, see b/441138434.

Bug: 440414803
Bug: 440280459
Test: Manual
Flag: com.android.window.flags.enable_display_disconnect_interaction
Change-Id: I859549425e88c77589b3fee69e68fb2c2ff6aa8f