Fix the bluetooth tile crash
The logic has been changed with dual shade. It was initilally a flag
that the details view feature depends on. But this logic has been
removed.

So currently when QsDetailedView is enabled but it's not in dual shade
mode, it will still show the dialog since the details feature is still
disabled. The DetailsViewModel checks if `isDualShade` and prevents
details from showing if it's not dual shade.

Bug: 440607916
Flag: com.android.systemui.qs_tile_detailed_view
Test: Manually
Change-Id: I5d781af990294d1777b4609d96d48c452b85c72f