[Flexiglass] Remove NotificationLockscreenScrim from LockscreenContent
NotificationLockscreenScrim was an opaque full screen box over the
Lockscreen, which had a matching background with the ShadeScene. This
provided a continuity in showing a black/white surface after the
ShadeScene -> LockscreenScene transition has finished, and finalized
this transition by starting a fade-out animation when composing the
LocksceenScene.

This transition broke when the ShadeScene became transparent, and
showing the NotificationLockscreenScrim resulted in a "flash".

This CL removes the "flash" by inverting the fade-out animation to get a
fade-in, and apply it to the Lockscreen content after the Shade ->
Lockscreen transition has finished.

Fixes: 437121774
Test: check the ShadeScene to LockscreenScene transition -> Lockscreen content fades in, the "flash" is gone
Flag: com.android.systemui.scene_container

Change-Id: I7b6581cca6da45ea9655e6acfbb34a7dbb071bbb