Hide keyguard at notifyBiometricAuthModeChanged to avoid flickering
When fingerprint sensor is located on power button the keyguard is
shown a short while before home screen when unlocking with fingerprint.

This only happens when all lock screen settings in "When to show"
are disabled. The reason is that updateIsKeyguard is not called
when all settings in "When to show" are disabled.

The solution is to call updateIsKeyguard already at
notifyBiometricAuthModeChanged.

Flag: com.android.systemui.update_keyguard_on_wake_and_unlock_earlier
Test: atest CentralSurfacesImplTest
Bug: 441381838
Change-Id: I2014cccb61a02faa0dce0487ed0b3355e138f2e5