Fix requesting suggestion after resuming the player.
- `visibleMediaPlayers` used to add a key when the player transitioned
from resumption state while keeping the old key associated with the
resumption state. After the change, the old key is replaced with the new
key.

- Set the `onSuggestionSpaceVisible` callback available to the UI even
if data.resumption is false. This allows the callback to be called when
the resumption player panel just transitioned to be active but haven't
received the new data.

Fix: 441358997
Test: atest MediaCarouselControllerTest, MediaDeviceManagerTest
Test: On a physical device
Flag: com.android.systemui.enable_suggested_device_ui
Change-Id: Ic9b2ef20b29306cf847c950dd42d8b676cd8ef29