Provide IME insets to freeform windows in the same task as IME target
After the change [1], for freeform windows, IME insets are provided only
when the target window is equals to the IME target.

However, we'd like to provide the IME insets as long as the target is in
the same task as the ime target window. This helps, for example,  a case
where non-input-focusable but ime-focusable dialog window monitors IME
insets.

[1] Id376f73b7c3b0315834e9be4eecc21a826db85fc

Bug: 440265933
Test: ImeInsetsVisibilityTest
Flag: EXEMPT bug fix
Change-Id: Ieb5d7b40e18453a54806bcc2239681c2ba69606c