Reduce flakes in MediaProjection flicker tests
MediaProjection flicker tests sometimes flake or outright fail due to a
StaleObjectException when trying to click on a popup menu. The exact
cause of this exception isn't known, possibly related to transitions,
but to resolve it, we are simply going to retry a set number of times to
see if it is possible to click the desired UI item. THis should resolve
the flakes, while still failing properly if something is actually wrong
with the scenario

Bug: 434632075
Bug: 435896354
Test: StartAppMediaProjectionResizeAndDrag
Test: StartAppMediaProjecitonWithMaxDesktopWindows
Test: StartScreenMediaProjectionWithMaxDesktopWindows
Flag: TEST_ONLY
Change-Id: Ic1280d96aaf66ff953b2c5728c829bb776a42cfa