Implement Clock Size Switch Animation
This is a first pass at the clock switch animation. It's functional
and animates the correct elements, but isn't to spec yet. An earlier
version used two seperate calls to NestedScenes for the clock size and
centering transitions. This simplified the scene transition graph, but
certain aspects of that approach didn't work.

Bug: 441339360
Bug: 438513876
Flag: com.android.systemui.scene_container
Test: Checked large / small clock switch animations
Change-Id: Ie86abb3fc79830f726cad24b207634685a2db4aa