A11y: announce when resolver items are "selected"
To test, I manually changed the condition in the listener (checking
layout-with-default, selection, and always button) to instead read
`if (true)` -- otherwise we would launch the target immediately upon
selection, w/o reading out the state text. Then I initiate a Resolver
session with Talkback on.

When an item is first tapped, Talkback adds a green frame around its
row and reads "double-tap to select / double-tap-and-hold to long
press." Double-tapping anywhere on the screen changes the background
color of that item and reads "selected." Single-tapping a different
item moves the green frame but not the background; single-tapping
back to the currently-selected item indicates that it's still in the
"selected" state. I believe this is the correct a11y behavior.

Fix: 399189789
Test: manually as described above; atest ResolverActivityTest
Flag: EXEMPT bugfix
Change-Id: Ic446a308df296a7c3cb15f3d100fa9c0fde7b73b