am: Rename getPackageList to getProcessPackageNames
The method `ProcessRecord#getPackageList()` returns all package names
that are running in a given process. The name `getPackageList` is
somewhat ambiguous.

This CL renames the method to `getProcessPackageNames()` to more
accurately reflect its behavior. The new name is more descriptive and
aligns better with the method's Javadoc.

This is a pure refactoring with no behavior changes.

Bug: 441412109
Test: m services.core
Flag: EXEMPT PURE_REFACTOR

Change-Id: I178fae56b60c6c21ea053af9ce0b631a07549b2c