Allow closing tasks even if not inside a desk
Freeform tasks may get created along with an app header without being
part of a desk. While this is always a bug in launch logic, if we happen
to get into this state we should allow the task to be closed using the X
button instead of throwing an illegal state exception.

Flag: EXEMPT bugfix
Fix: 442865973
Test: atest DesktopTasksControllerTest
Change-Id: Ie62261ae54c6e785f2fda5d3bf19936d0d990cd9