Remove queued before abort during failsafe flush
It may be possible that a legacy sync's abort runs
logic that can allow other logic to run which may read
from mQueuedTransitions before we actually clear it.
So, to be safe, clear the queue first.

Also, add some more checks for isFlushing to ensure we
aren't accidentally adding to the queue or collection
during the flush.

Bug: 439829314
Test: Existing tests
Flag: EXEMPT bugfix
Change-Id: I82d5cb20d843184c40b6acf98d2167fef0880893