Read and write the cross-platform manifest for cross-platform transfers
On backup we write the cross-platform manifest to the full backup tar
file so it can be sent to the target device. On restore we check whether
the locally installed app has opted-in to cross-platform transfers the
the configured bundle id and team id matches the values in the manifest
that we received from the source device.

Bug: 432673356
Test: atest FullRestoreEngineTest.java
Test: Manual testing with bmgr backupnow/restore
Flag: com.android.server.backup.enable_cross_platform_transfer
Change-Id: I60b4435828cd0fc876cbe883b23deefd72aa7117